+++ tools/cli/embraion/__init__.py
@@ -1,3 +1,3 @@
 """EmbrAIon CLI package."""
 
-__version__ = "0.2.2"
+__version__ = "0.3.0.dev0"